src/requirements.txt
lxml==4.7.1
numpy==1.22.1
pandas==1.3.5
Pillow==9.0.0
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2021.3
six==1.16.0
